# Optimize overview page rendering by only acting on latest jobs